[Refact] Extrair Classe Origem: FileReader e FileWriter Impactos: Persistencia.java Explicação: os métodos de leitura dos arquivos de entrada e escrita de arquivo de saída foram movidos para uma nova classe chamada Persistência.java
